+++ src/kcii_audit/parsers/__init__.py
@@ -3,11 +3,13 @@ from kcii_audit.parsers.dbms_mysql import records_from_mysql_paste
 from kcii_audit.parsers.dbms_postgresql import records_from_postgresql_paste
 from kcii_audit.parsers.linux_server import records_from_linux_paste
 from kcii_audit.parsers.network_cisco_ios import records_from_cisco_ios_paste
+from kcii_audit.parsers.network_junos import records_from_junos_paste
 from kcii_audit.parsers.unix_server import records_from_unix_paste
 from kcii_audit.parsers.windows_server import records_from_windows_paste
 
 __all__ = [
     "records_from_cisco_ios_paste",
+    "records_from_junos_paste",
     "records_from_linux_paste",
     "records_from_mariadb_paste",
     "records_from_mysql_paste",